requirements.txt
# AntiGravity System Dependencies

# Core dependencies
pyyaml==6.0.1
python-dotenv==1.0.0

# Trading platforms (optional - install as needed)
# binance-connector==3.7.0
# pybit==5.6.2
# MetaTrader5==5.0.45

# Vision and automation (from existing agent)
opencv-python==4.8.1.78
pillow==10.1.0
pyautogui==0.9.54
pytesseract==0.3.10
numpy==1.26.2

# Audio capture (from existing agent)
sounddevice==0.4.6
soundfile==0.12.1

# GUI (from existing agent)
customtkinter==5.2.1

# AI/LLM integration
# google-generativeai==0.3.1  # Uncomment if using Gemini API

# Utilities
requests==2.31.0
